refactor(AdditionalInfoScreen): mover estilos para arquivo styles.js
