+++ mavis/src/AdditionalInfoScreen/styles.js
@@ -1,0 +1,28 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 30,
+    justifyContent: 'center',
+    backgroundColor: '#fff',
+  },
+  title: {
+    textAlign: 'center',
+    marginBottom: 20,
+  },
+  subtitle: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    marginVertical: 10,
+  },
+  input: {
+    marginBottom: 6,
+  },
+  button: {
+    marginTop: 10,
+    alignSelf: 'center',
+    width: '100%',
+  },
+});
+
+export default styles
